requirements.txt
playwright==1.42.0
pytest==8.0.2
pytest-playwright==0.4.4
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==1.0.1
allure-pytest==2.13.2 